requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
aiosqlite==0.20.0
python-dotenv==1.0.1
APScheduler==3.10.4
pytz==2025.1
pillow==10.4.0
